+++ CHANGELOG.md
@@ -2,7 +2,9 @@
 ## v1.2.0
 + Added new value type float.
 + Hexadecimal values can be prefixed with 0x (unless the values of HEX:).
-
++ Variables.
++ Include directive.
++ Macros.
 ## v1.1.0
 + Added API for autocompletion.
 + Added python parser.